+++ appium-dotnet-driver/Appium/Interfaces/IFindsByFluentSelector.cs
@@ -11,24 +11,29 @@
 //WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //See the License for the specific language governing permissions and
 //limitations under the License.
+
 using System.Collections.ObjectModel;
 
 namespace OpenQA.Selenium.Appium.Interfaces
 {
-    public interface IFindByAccessibilityId<W> : IFindsByFluentSelector<W> where W : IWebElement
+    public interface IFindsByFluentSelector<W> where W : IWebElement
     {
         /// <summary>
-        /// Finds the first of elements that match the Accessibility Id selector supplied
+        /// Method performs the searching for a single element by some selector defined by string
+        /// and value of the given selector
         /// </summary>
-        /// <param name="selector">an Accessibility Id selector</param>
-        /// <returns>IWebElement object so that you can interact that object</returns>
-        W FindElementByAccessibilityId(string selector);
+        /// <param name="by">is a string selector</param>
+        /// <param name="value">is a value of the given selector</param>
+        /// <returns>the first found element</returns>
+        W FindElement(string by, string value);
 
         /// <summary>
-        /// Finds a list of elements that match the Accessibility Id selector supplied
+        /// Method performs the searching for a list of elements by some selector defined by string
+        /// and value of the given selector
         /// </summary>
-        /// <param name="selector">an Accessibility Id selector</param>
-        /// <returns>IWebElement object so that you can interact that object</returns>
-        ReadOnlyCollection<W> FindElementsByAccessibilityId(string selector);
+        /// <param name="by">is a string selector</param>
+        /// <param name="value">is a value of the given selector</param>
+        /// <returns>a list of elements</returns>
+        ReadOnlyCollection<W> FindElements(string selector, string value);
     }
 }